feat: Export types for Resource
