+++ packages/experimental/src/index.ts
@@ -4,3 +4,4 @@ export { default as Entity } from './entity/Entity';
 export { default as EntityRecord } from './entity/EntityRecord';
 export { default as Resource } from './rest/Resource';
 export { default as BaseResource } from './rest/BaseResource';
+export * from './rest/types';
